+++ robospice-cache-parent/robospice-cache-test/src/main/java/com/octo/android/robospice/persistence/keysanitizer/AbstractKeySanitizerTest.java
@@ -1,0 +1,36 @@
+package com.octo.android.robospice.persistence.keysanitizer;
+
+import android.test.AndroidTestCase;
+
+import com.octo.android.robospice.persistence.keysanitation.KeySanitizer;
+
+public abstract class AbstractKeySanitizerTest extends AndroidTestCase {
+
+    protected KeySanitizer keySanitizer;
+
+    protected void setUp(KeySanitizer keySanitizer) throws Exception {
+        super.setUp();
+        this.keySanitizer = keySanitizer;
+    }
+
+    public void testSanitizeKey(Object expectedSanitizedKey, Object cacheKeyToSanitize) {
+        // given
+
+        // when
+        Object sanitized = keySanitizer.sanitizeKey(cacheKeyToSanitize);
+
+        // then
+        assertEquals(expectedSanitizedKey, sanitized);
+    }
+
+    public void testDesanitizeKey(Object expectedDesanitizedKey, Object cacheKeyToDesanitize) {
+        // given
+
+        // when
+        Object sanitized = keySanitizer.desanitizeKey(cacheKeyToDesanitize);
+
+        // then
+        assertEquals(expectedDesanitizedKey, sanitized);
+    }
+
+}